AI-Code-Coach-master/requirements.txt
langchain
langchain-groq
faiss-cpu
sentence-transformers
streamlit
python-dotenv